Fix: 끝전 관련 자료형 BigDecimal 통일, 버그 수정

+++ ledger-api/src/main/java/com/pocketstock/ledger/cma/domain/CollectionSetting.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -18,7 +19,7 @@ public class CollectionSetting {
     private String sourceType;      // ACCOUNT / CARD / POINT
     private Long sourceRefId;       // linked_account_id 등
     private Boolean isEnabled;
-    private Integer threshold;           // 끝전 커팅 기준: 1000 / 5000 / 10000 (ACCOUNT 타입에만 적용)
+    private BigDecimal threshold;        // 끝전 커팅 기준: 1000 / 5000 / 10000 (ACCOUNT 타입에만 적용)
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
